+++ hydra-node/exe/hydra-node/Main.hs
@@ -4,13 +4,21 @@ module Main where
 
 import Hydra.Prelude
 
+import Crypto.Random (getRandomBytes)
 import Hydra.API.Server (Server (Server, sendOutput), withAPIServer)
 import Hydra.API.ServerOutput (ServerOutput (PeerConnected, PeerDisconnected))
-import Hydra.Cardano.Api (serialiseToRawBytesHex)
+import Hydra.Cardano.Api (
+  File (..),
+  Key (SigningKey),
+  getVerificationKey,
+  serialiseToRawBytesHex,
+  writeFileTextEnvelope,
+ )
 import Hydra.Chain.CardanoClient (QueryPoint (..), queryGenesisParameters)
 import Hydra.Chain.Direct (initialChainState, loadChainContext, mkTinyWallet, withDirectChain)
 import Hydra.Chain.Direct.ScriptRegistry (publishHydraScripts)
 import Hydra.Chain.Direct.Util (readKeyPair)
+import Hydra.Crypto (HydraKey, generateSigningKey)
 import Hydra.HeadLogic (
   Environment (..),
   Event (..),
@@ -43,7 +51,8 @@ import Hydra.Node (
 import Hydra.Node.EventQueue (EventQueue (..), createEventQueue)
 import Hydra.Options (
   ChainConfig (..),
-  Command (Publish, Run),
+  Command (GenHydraKey, Publish, Run),
+  GenerateKeyPair (GenerateKeyPair, outputFile),
   LedgerConfig (..),
   PublishOptions (..),
   RunOptions (..),
@@ -52,6 +61,7 @@ import Hydra.Options (
   validateRunOptions,
  )
 import Hydra.Persistence (createPersistenceIncremental)
+import System.FilePath ((<.>))
 
 main :: IO ()
 main = do
@@ -62,6 +72,8 @@ main = do
       run (identifyNode options)
     Publish options ->
       publish options
+    GenHydraKey outputFile ->
+      genHydraKeys outputFile
  where
   run opts = do
     let RunOptions{verbosity, monitoringPort, persistenceDir} = opts
@@ -107,6 +119,11 @@ main = do
     let PublishOptions{publishNetworkId = networkId, publishNodeSocket} = opts
     txId <- publishHydraScripts networkId publishNodeSocket sk
     putStr (decodeUtf8 (serialiseToRawBytesHex txId))
+
+  genHydraKeys GenerateKeyPair{outputFile} = do
+    sk :: SigningKey HydraKey <- generateSigningKey <$> getRandomBytes 16
+    void $ writeFileTextEnvelope (File (outputFile <.> "sk")) Nothing sk
+    void $ writeFileTextEnvelope (File (outputFile <.> "vk")) Nothing (getVerificationKey sk)
 
   withNetwork tracer Server{sendOutput} signingKey parties host port peers nodeId =
     let localhost = Host{hostname = show host, port}